resort_api/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
cachetools==6.2.1
certifi==2025.10.5
click==8.3.0
fastapi==0.119.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.1.0
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
pydantic==2.12.1
pydantic_core==2.41.3
Pygments==2.19.2
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-mock==3.15.1
python-dotenv==1.1.1
PyYAML==6.0.3
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
